% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_query_url.R
\name{construct_query_url}
\alias{construct_query_url}
\title{Construct URL for API query}
\usage{
construct_query_url(req_defaults, api_key = Sys.getenv("PVWATTS_KEY"))
}
\arguments{
\item{req_defaults}{Required parameters for API request}

\item{api_key}{API key. Default is Sys.getenv("PVWATTS_KEY")}
}
\value{
req_url URL for API request
}
\description{
Construct URL for API query
}
